fix :: deleteMemberQuery #100
