+++ src/components/Group/GroupDetail/Modal/memberInfoModal.tsx
@@ -3,35 +3,15 @@ import React from 'react';
 import * as S from '../style';
 import { GroupMember } from 'repositories/Group/group.repository';
 import { groupPermissionTransform } from 'utils/Transform/gruopPermissionTransform';
-import { usePatchGroupMemberStatusMutation } from 'queries/Group/group.query';
-import { B1ndToast } from '@b1nd/b1nd-toastify';
 
 interface MemberInfoModalProps {
   isOpen: boolean;
   onClose: () => void;
   member: GroupMember;
+  patchGroupMemberStatus: () => void;
 }
 
-const MemberInfoModal = ({ isOpen, onClose, member }: MemberInfoModalProps) => {
-  const patchGroupMemberStatusMutation = usePatchGroupMemberStatusMutation();
-  const patchGroupMemberStatus = () => {
-    patchGroupMemberStatusMutation.mutate(
-      {
-        status: 'REJECTED',
-        id: member.id,
-        memberId: [member.id],
-      },
-      {
-        onSuccess: () => {
-          B1ndToast.showSuccess('내보내기 성공');
-          onClose();
-        },
-        onError: () => {
-          B1ndToast.showError('내보내기 실패');
-        },
-      }
-    );
-  };
+const MemberInfoModal = ({ isOpen, onClose, member, patchGroupMemberStatus }: MemberInfoModalProps) => {
   return (
     <DodamModal isOpen={isOpen} close={onClose} customStyle={{ top: '-1%' }}>
       <S.MemberInfoModalWrap>
